gcode/requirements.txt
Pillow==4.1.1
Flask==0.12.2

